`driver`: better test framework (#1348)
Progress on https://github.com/cowprotocol/services/issues/1306.

Refactor the test framework to make tests easy to write. Refactor some
of the tests to use the new test framework, which is in `setup::new`.
I'm aware of all TODOs and I'm leaving them for follow-up PRs.
Eventually I plan to migrate the `setup::new` module to be simply
`setup`.