Fix Divide By Zero Panic (#1439)
This PR fixes a panic that was happening because of a "divide by 0" in
the single order solver prioritization logic. It does so in two ways:

- First, it adjusts the code to no longer panic on 0 amounts. In theory,
the `native_buy_amount` should never really be 0, but dealing with the
case is very simple, so I decided to handle it defensively.
- Second, we now discard partially fillable orders where one of the
amounts scale to 0. This can happen in cases where the target amount
(i.e. buy amount for sell orders, sell amount for buy orders) has a
smaller value than the side amount, and the order is mostly filled. Most
of our code assumes orders without 0 amounts, so we should not produce
limit orders with these amounts anywhere.

### Test Plan

Added unit tests to verify logic changes.