Fix Order Execution to Trade Matching (#1438)
Fixes #1437 

@fhenneke highlighted some discrepancies in the computed observed fees
and the actual solver fees for any given settlement.

This is the result of two issues:
- the `executedAmount` encoded in a trade in a settlement transaction
**gets ignored** for fill or kill orders. As such, it is not correct to
compare these amounts when matching orders. This shouldn't be an issue
with the fee `fold` logic, the `solver_fee` amount is the same for each
FoK order execution (for which there should be exactly 1).
- The computed total `executedAmount` for an order was using the
`executedSellAmountWithFees` and not the simple `executedSellAmount`, as
such it was computing an incorrect total executed amount for sell
orders.

This PR addresses those two issues.

### Test Plan

Added a unit tests to verify that the example settlement from the [Slack
thread](https://cowservices.slack.com/archives/C0361CDD1FZ/p1681824222315879)
would correctly compute fees.