Adjust score calculator for reverts (#1458)
Fixes https://github.com/cowprotocol/services/issues/1453

I thought about putting the change under the configuration parameter,
but it seems like it's not needed since we are going to use MEVBlocker
permanently. Also, new parameter would need to be coupled somehow with
the transactions submission strategy parameters (every time we need to
update the submission configuration we would potentially need to also
remember to update the configuration for score calculator).

@harisang can you double check that the unit test change is correct? Now
the score is slightly increased.