Always capture backtrace on panic (#1467)
`Backtrace::capture` doesn't capture a backtrace if the environment
variable RUST_LIB_BACKTRACE is set to 0 because capturing a backtrace is
slow. This PR switches to `force_captures` which does what the name
implies. This is fine because this code isn't performance sensitive
because it only runs just before killing the process after a panic
anyway.