CI improvements (#1503)
Fixes https://github.com/cowprotocol/services/issues/1105.

Add driver tests to CI. Start using `docker-compose` in CI instead of
translating the compose setup into a GitHub workflow. This has a few
benefits: it ensures that the local environment setup is on-par with the
CI environment setup, it removes the burden of maintaining two different
environment setup flows, and it tests the docker-compose setup in CI so
it's impossible to commit code with a broken docker-compose.

There is one issue: because dev-geth is part of this repo, it needs to
be built from scratch every time the dev-geth compose service is
started. This adds up to 10 minutes to the CI pipeline. The solution to
this problem is to extract dev-geth into its own repository, with its
own release schedule and push the docker image to the registry. This way
the dev-geth code will not need to get rebuilt all the time. I opened
[an issue](https://github.com/cowprotocol/services/issues/1505) for
this.