Remove app data hash based fee subsidy (#1528)
Related to https://github.com/cowprotocol/services/issues/1465 . With
the meaning of app data changing, it doesn't make sense to award a
subsidy to an arbitrary app data hash value because the value is a hash
of something else and not freely chooseable.

If we wanted to keep app id (which is not app data hash) based subsidies
then we would need to do that through the full app data. But this
feature in general makes little sense because anyone could pick an app
id that has a subsidy even when they're not actually coming from that
app. So it makes more sense to remove. I don't think we're actively
using this anyway. In the infra configuration I see one value which has
last been mentioned on Slack more than a year ago.