Create app-data-hash crate (#1532)
Part of https://github.com/cowprotocol/services/issues/1465 .

Note that this crate doesn't concern itself with the full app data
format. It treats it as raw bytes. I've done it this way to make it easy
to reuse. We will have more specific validation logic in the order
creation code.

I am using the tiny-keccak crate because we already depend on it through
web3.