Refactor OrderCreation to allow for new app data scheme (#1556)
In
https://github.com/cowprotocol/services/pull/1537#discussion_r1212850311
we discussed how the new app data format should be added in a backward
compatible way. This PR implements Nick's suggestion:

> My suggestion would be to have `appData` be either a "full" app data
(where a sidecar `app_data_hash` can be specified to assist with
validation) or a hash. So all of these would be valid:
> 
> ```js
> { "appData": "0x...." }
> { "appData": "{\"hello\":\"world\"}", "appDataHash": "0x...." }
> { "appData": "{\"hello\":\"world\"}" }
> ```

In order to implement the new format I made the following changes:
- Stop storing OrderData in OrderCreation. It is no longer true that all
fields from OrderData exist in the same way in OrderCreation. In
OrderData app data is always a hash. In OrderCreation app data can be in
several forms. We can go from OrderCreation to OrderData but not the
other way around.
- Remove conversions between OrderCreation and Order. It is unclear how
they should work because they need to make guesses about some fields. It
is better to do this explicitly. For an example of this see
`crates/shared/src/order_validation.rs` line 575, which requires very
specific values that weren't documented in `Order::from_order_creation`.
- Derive Default for OrderCreation. If you need a valid signature, use
`.sign`.
- Temporarily error in post_order if the new app data is used.