Add solver computation count metric (#1592)
I noticed we have a Grafana panel "Solver Solution Computation Time"
which tries to get an average of solver computation time by using the
prometheus function `count_over_time`. This isn't right because
depending on Prometheus' sampling interval it is impossible to get the
true count. This PR adds the count.

We should consider making this a histogram but I wanted to make this
simpler change immediately.