Fetch full app data from IPFS when not specified (#1587)
We want to continue to support IPFS based app data even when on order
creation only the contract app data hash is specified. This PR changes
how orders with only a contract app data hash are handled.

Before:
Order is accepted and we may or may not already have corresponding full
app data in the database.

Now:
First try to find existing app data in database. If it doesn't exist try
to find it on IPFS with the old and new method of CID generation. If it
still doesn't exist then continue as before without full app data. If it
does exist, then don't try to verify the hash of the found full app data
because this might be the old way of storing on IPFS, which we cannot
validate. Or it might be hardcoded app data that doesn't correspond to
any IPFS data. Do verify that the schema is as expected.

The last step is a breaking change. Before this PR we would accept an
order that has a contract app data hash that resolves to an IPFS
document that isn't json (or not utf-8). We would accept it because we
didn't have the full app data to validate. With this PR, when we find
the full app data, we will validate the schema.