`driver`: encode auction ID instead of settlement ID (#1634)
Encode the auction ID in the settlement calldata instead of the
settlement ID. Get rid of the concept of settlement IDs. Update the
`/settle` endpoint to not require any parameters. In the autopilot, wait
for a settlement with the auction ID to be published by the winning
address.

### Test Plan

Automated tests. Eventually will be tested on staging as well, after
other driver infra bugs are handled.