Filter Out Invalid Uniswap V2 Pools (#1639)
This PR fixes an issue where the `driver` was sending invalid UniV2
pools to the solver engine. The engine expects ~~`uint96`~~ `uint112`
reserves, but it is possible to set up a pools that has more reserves
than that (effectively bricking it).

### Test Plan

Ran locally on Goerli and swapped with WETH <>
`0xdc31ee1784292379fbb2964b3b9c4124d8f89c60` pool.