Don't track cancellation tx when submission stops on simulation revert (#1749)
Follow up fix for #1723 and #1544

Both of these fixes focussed on not tracking the cancellation
transaction **when the deadline was reached**. However, we also cancel
transactions when the simulation starts reverting, cf:

https://github.com/cowprotocol/services/blob/96b039be1b5eb8bf2a3cea39591553faa3d6ee2c/crates/solver/src/settlement_submission/submitter.rs#L460-L475

In this case we don't remove the pending cancellation transaction (which
will never get mined) and indeed persist it for future tracking (which
will prevent further runloops of that solver if the base fee decreases).

I believe the proper fix is to **never persist** any pending transaction
when soft cancellations are turned on. We will always either have a
mined transaction or submit a cancellation tx which means in the case of
soft cancellations there is never the need to track anything in the next
runloop.

Note, that it is fine to persist empty transaction lists (as this would
mean nothing needs to be tracked), however we still need to wait for the
original settlement transaction to potentially propagate (in case our
soft cancellation raced with on-chain inclusion). I believe this wasn't
correct in the old implementation either (where we may skip "waiting"
for the mined tx in case the deadline was reached).

### Test Plan

This piece of code is extremely hard to reason about and poorly unit
tested. My plan is to observe this change in barn before it gets
released into prod. We should seriously consider refactoring this piece
of code.